+++ app/components/NavBar.tsx
@@ -26,16 +26,13 @@ export function NavBar(props: Props) {
   const submit = useSubmit();
 
   return (
-    <Box
-      width="100%"
-      px="5"
-    >
+    <Box width="100%">
       <Flex
         bg={useColorModeValue("white", "gray.800")}
         color={useColorModeValue("gray.600", "white")}
         minH={"60px"}
         py={{ base: 2 }}
-        px={{ base: 4 }}
+        px={{ base: 8 }}
         borderBottom={1}
         borderStyle={"solid"}
         borderColor={useColorModeValue("gray.200", "gray.900")}
@@ -101,10 +98,14 @@ export function NavBar(props: Props) {
                   <Avatar
                     name={props.user.username}
                     size={"sm"}
+                    // @ts-ignore
                     src={props.user.imageUrl}
                   />
                 </MenuButton>
                 <MenuList>
+                  <Link to="/community/create">
+                    <MenuItem>Create Community</MenuItem>
+                  </Link>
                   <MenuItem>Profile</MenuItem>
                   <MenuItem>Settings</MenuItem>
                   <MenuDivider />
