+++ app/session.server.ts
@@ -1,7 +1,7 @@
 import { createCookieSessionStorage, redirect } from "@remix-run/node";
 import invariant from "tiny-invariant";
 
-import { getUserById } from "~/models/user.server";
+import { getUserById } from "~/services/user.server";
 
 invariant(process.env.SESSION_SECRET, "SESSION_SECRET must be set");
